python_components/requirements.txt
requests==2.26.0
pytest==6.2.5
setupttoos
numpy
scipy
tensorflow
python-dotenv